refactor(dbt): move `DbtManifestAssetSelection` to `resources_v2`
