+++ python_modules/libraries/dagster-dbt/dagster_dbt_tests/cli/test_asset_selection.py
@@ -7,7 +7,7 @@ from dagster._utils import file_relative_path
 from dagster_dbt.asset_decorator import dbt_assets
 from dagster_dbt.cli.resources_v2 import DbtManifest
 
-manifest_path = file_relative_path(__file__, "sample_manifest.json")
+manifest_path = file_relative_path(__file__, "../sample_manifest.json")
 manifest = DbtManifest.read(path=manifest_path)
